@@ -1,8 +1,9 @@
 {
   "name": "express-api-starter",
   "version": "1.2.0",
-  "description": "A basic starter for an express.js API",
+  "description": "A basic starter for an express.js API & Vonage Cloud Runtime (VCR)",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "node src/index.js",
     "dev": "nodemon src/index.js",
@@ -10,25 +11,27 @@
     "test": "jest"
   },
   "keywords": [],
-  "author": "CJ R. <cj@null.computer> (https://w3cj.now.sh)",
+  "author": "Toni Kuschan <toni.kuschan@vonage.com>",
   "repository": {
     "type": "git",
-    "url": "https://github.com/w3cj/express-api-starter.git"
+    "url": "https://github.com/klapperkopp/express-api-starter-vcr.git"
   },
   "license": "MIT",
   "dependencies": {
+    "@vonage/jwt": "^1.11.0",
+    "@vonage/server-sdk": "^3.19.4",
+    "@vonage/vcr-sdk": "^1.3.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.21.1",
+    "dotenv": "^16.4.7",
+    "express": "^4.21.2",
     "helmet": "^8.0.0",
     "morgan": "^1.10.0"
   },
   "devDependencies": {
-    "eslint": "^8.57.0",
-    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint": "^9.20.1",
     "eslint-plugin-import": "^2.31.0",
     "jest": "^29.7.0",
-    "nodemon": "^3.1.7",
+    "nodemon": "^3.1.9",
     "supertest": "^7.0.0"
   }
 }
