@@ -1,4 +1,4 @@
-const app = require('./app');
+import app from './app.js';
 
 const port = process.env.PORT || 5000;
 app.listen(port, () => {
